Restructured examples.
