+++ examples/core-api/scan/app.js
