+++ examples/core-api/scan/evothings.json
@@ -1,0 +1,3 @@
+{
+  "uuid": "7cc640aa-b54c-4ae0-a78b-e6f07961465b"
+}
