+++ examples/core-api/tisensortag-low-level-api/evothings.json
@@ -1,0 +1,3 @@
+{
+  "uuid": "83881f3d-20fa-4129-9429-b299d5faf3c7"
+}
