+++ examples/core-api/tisensortag-low-level-api/index.html
@@ -2,7 +2,7 @@
 <html>
 
 <!--
-Demo of using the BLE Plugin API for reading the luxometer of the TI SensorTag CC2650.
+Demo of using the old low-level BLE Plugin API for reading the luxometer of the TI SensorTag CC2650.
 -->
 
 <head>
@@ -10,7 +10,7 @@ Demo of using the BLE Plugin API for reading the luxometer of the TI SensorTag C
 	<meta name="viewport" content="width=device-width, user-scalable=no,
 		shrink-to-fit=no, initial-scale=1.0, minimum-scale=1.0, maximum-scale=1.0" />
 
-	<title>TI SensorTag Demo - BLE Plugin API</title>
+	<title>TI SensorTag Demo - Low-Level BLE Plugin API</title>
 
 	<script>
 	// Redirect console.log when running from Evothings Workbench.
@@ -25,9 +25,9 @@ Demo of using the BLE Plugin API for reading the luxometer of the TI SensorTag C
 <body>
 	<img class="evo-image" src="https://evomedia.evothings.com/2013/11/evothings_logo_large_white1.png" />
 
-	<h1>TI SensorTag Demo - BLE Plugin API</h1>
+	<h1>TI SensorTag Demo - Low-Level BLE Plugin API</h1>
 
-	<p>This app demonstrates how to use the BLE plugin API for reading the luxometer on the TI SensorTag CC2650.</p>
+	<p>This app demonstrates how to use the old low-Level BLE plugin API for reading the luxometer on the TI SensorTag CC2650.</p>
 
 	<p>Activate the SensorTag. If you don't have a SensorTag you can modify the code in file <code>app.js</code> to work with your device.</p>
 
@@ -42,16 +42,9 @@ Demo of using the BLE Plugin API for reading the luxometer of the TI SensorTag C
 	<script src="cordova.js"></script>
 
 	<!--
-	Include the extended BLE API.
-	-->
-	<script src="extended-ble.js"></script>
-
-	<!--
 	Include application code.
 	-->
-	<script src="app-using-higher-level-api-with-automatic-service-discovery.js"></script>
-	<!--<script src="app-using-higher-level-api.js"></script>-->
-	<!--<script src="app-using-lower-level-api.js"></script>-->
+	<script src="app.js"></script>
 </body>
 
 </html>
