+++ examples/core-api/tisensortag/app.css
@@ -7,13 +7,13 @@ body
 
 h1
 {
-	font-size: 250%;
-	margin-right: 110px;
+	font-size: 150%;
+	margin-right: 120px;
 }
 
 p
 {
-	font-size: 130%;
+	font-size: 110%;
 }
 
 #message
