+++ examples/core-api/tisensortag/evothings.json
@@ -1,0 +1,3 @@
+{
+  "uuid": "43fde818-b69c-48c7-8f43-7d91f3254899"
+}
