+++ examples/core-api/tisensortag/index.html
@@ -2,7 +2,7 @@
 <html>
 
 <!--
-Demo of using the EasyBLE library for reading the luxometer of the TI SensorTag CC2650.
+Demo of using the high-level BLE Plugin API for reading the luxometer of the TI SensorTag CC2650.
 -->
 
 <head>
@@ -10,7 +10,7 @@ Demo of using the EasyBLE library for reading the luxometer of the TI SensorTag 
 	<meta name="viewport" content="width=device-width, user-scalable=no,
 		shrink-to-fit=no, initial-scale=1.0, minimum-scale=1.0, maximum-scale=1.0" />
 
-	<title>TI SensorTag Demo - EasyBLE</title>
+	<title>TI SensorTag Demo - BLE Plugin API</title>
 
 	<script>
 	// Redirect console.log when running from Evothings Workbench.
@@ -25,9 +25,9 @@ Demo of using the EasyBLE library for reading the luxometer of the TI SensorTag 
 <body>
 	<img class="evo-image" src="https://evomedia.evothings.com/2013/11/evothings_logo_large_white1.png" />
 
-	<h1>TI SensorTag Demo - EasyBLE</h1>
+	<h1>TI SensorTag Demo - BLE Plugin API</h1>
 
-	<p>This app demonstrates how to use the EasyBLE library for reading the luxometer on the TI SensorTag CC2650.</p>
+	<p>This app demonstrates how to use the BLE plugin API for reading the luxometer on the TI SensorTag CC2650.</p>
 
 	<p>Activate the SensorTag. If you don't have a SensorTag you can modify the code in file <code>app.js</code> to work with your device.</p>
 
@@ -40,11 +40,6 @@ Demo of using the EasyBLE library for reading the luxometer of the TI SensorTag 
 	for the Cordova BLE plugin.
 	-->
 	<script src="cordova.js"></script>
-
-	<!--
-	Include the EasyBLE library.
-	-->
-	<script src="easyble.dist.js"></script>
 
 	<!--
 	Include application code.
