--- examples/core/scan/evothings.json
@@ -1,3 +1,0 @@
-{
-  "uuid": "0066d575-6cd2-453c-9644-b5fbb6341144"
-}
