--- examples/core/tisensortag/evothings.json
@@ -1,3 +1,0 @@
-{
-  "uuid": "d9fb811e-f547-4e84-a986-f73b459b3442"
-}
