+++ examples/deprecated/easyble/tisensortag/app.js
@@ -1,4 +1,5 @@
-// Demo of extended BLE plugin API.
+// Demo of the EasyBLE library (this library is deprecated,
+// use the updated BLE plugin API which uses a high-level style).
 
 // Application code starts here. The code is wrapped in a
 // function closure to prevent overwriting global objects.
@@ -12,9 +13,6 @@ var LUXOMETER_DATA = 'f000aa71-0451-4000-b000-000000000000'
 
 function initialize()
 {
-	// Extend plugin API with new functions (see file new-api.js).
-	extendBLEPluginAPI()
-
 	// Start scanning for a device.
 	findDevice()
 }
@@ -24,20 +22,20 @@ function findDevice()
 	showMessage('Scanning for the TI SensorTag CC2650...')
 
 	// Start scanning. Two callback functions are specified.
-	evothings.ble.startScan(
-		onDeviceFound,
-		onScanError,
+	evothings.easyble.startScan(
+		deviceFound,
+		scanError,
 		{ serviceUUIDs: ['0000aa10-0000-1000-8000-00805f9b34fb'] })
 
 	// This function is called when a device is detected, here
 	// we check if we found the device we are looking for.
-	function onDeviceFound(device)
+	function deviceFound(device)
 	{
 		// For debugging, print advertisement data.
 		//console.log(JSON.stringify(device.advertisementData))
 
-		// Alternative way to identify the device by advertised service UUID.
 		/*
+		// Alternative way to identify the device by advertised service UUID.
 		if (device.advertisementData.kCBAdvDataServiceUUIDs.indexOf(
 			'0000aa10-0000-1000-8000-00805f9b34fb') > -1)
 		{
@@ -45,12 +43,12 @@ function findDevice()
 		}
 		*/
 
-		if (device.advertisementData.kCBAdvDataLocalName == 'CC2650 SensorTag')
+		if (device.getName() == 'CC2650 SensorTag')
 		{
 			showMessage('Found the TI SensorTag!')
 
 			// Stop scanning.
-			evothings.ble.stopScan()
+			evothings.easyble.stopScan()
 
 			// Connect.
 			connectToDevice(device)
@@ -58,7 +56,7 @@ function findDevice()
 	}
 
 	// Function called when a scan error occurs.
-	function onScanError(error)
+	function scanError(error)
 	{
 		showMessage('Scan error: ' + error)
 	}
@@ -66,78 +64,84 @@ function findDevice()
 
 function connectToDevice(device)
 {
-	showMessage('Connecting to device...')
+	device.connect(connectSuccess, connectError)
 
-	evothings.ble.connectToDevice(
-		device,
-		onConnected,
-		onDisconnected,
-		onConnectError,
-		{ serviceUUIDs: [LUXOMETER_SERVICE] })
-
-	function onConnected(device)
+	function connectSuccess(device)
 	{
-		showMessage('Connected')
+		showMessage('Connected to device, reading services...')
 
-		// Get Luxometer service and characteristics.
-		var service = evothings.ble.getService(device, LUXOMETER_SERVICE)
-		var configCharacteristic = evothings.ble.getCharacteristic(service, LUXOMETER_CONFIG)
-		var dataCharacteristic = evothings.ble.getCharacteristic(service, LUXOMETER_DATA)
-
-		// Enable notifications for Luxometer.
-		enableLuxometerNotifications(device, configCharacteristic, dataCharacteristic)
+		// Read all services, characteristics and descriptors.
+		device.readServices(
+			readServicesSuccess,
+			readServicesError,
+			{ serviceUUIDs: [LUXOMETER_SERVICE] })
 	}
 
-	function onDisconnected(device)
+	function readServicesSuccess(device)
 	{
-		showMessage('Device disconnected')
+		showMessage('Reading services completed')
+
+		// Enable notifications for Luxometer.
+		enableLuxometerNotifications(device)
+	}
+
+	function readServicesError(error)
+	{
+		showMessage('Read services error: ' + error)
 	}
 
 	// Function called when a connect error or disconnect occurs.
-	function onConnectError(error)
+	function connectError(error)
 	{
-		showMessage('Connect error: ' + error)
+		if (error == evothings.easyble.error.DISCONNECTED)
+		{
+			showMessage('Device disconnected')
+		}
+		else
+		{
+			showMessage('Connect error: ' + error)
+		}
 	}
 }
 
 // Use notifications to get the luxometer value.
-function enableLuxometerNotifications(device, configCharacteristic, dataCharacteristic)
+function enableLuxometerNotifications(device)
 {
 	// Turn Luxometer ON.
-	evothings.ble.writeCharacteristic(
-		device,
-		configCharacteristic,
+	device.writeCharacteristic(
+		LUXOMETER_SERVICE,
+		LUXOMETER_CONFIG,
 		new Uint8Array([1]),
-		onLuxometerActivated,
-		onLuxometerActivatedError)
+		turnOnLuxometerSuccess,
+		turnOnLuxometerError)
 
-	function onLuxometerActivated()
+	// Enable notifications from the Luxometer.
+	device.enableNotification(
+		LUXOMETER_SERVICE,
+		LUXOMETER_DATA,
+		readLuxometerSuccess,
+		readLuxometerError)
+
+	function turnOnLuxometerSuccess()
 	{
 		showMessage('Luxometer is ON')
 
-		// Enable notifications from the Luxometer.
-		evothings.ble.enableNotification(
-			device,
-			dataCharacteristic,
-			onLuxometerNotification,
-			onLuxometerNotificationError)
 	}
-
-	function onLuxometerActivatedError(error)
+	function turnOnLuxometerError(error)
 	{
-		showMessage('Luxometer activate error: ' + error)
+		showMessage('Write Luxometer error: ' + error)
 	}
 
 	// Called repeatedly until disableNotification is called.
-	function onLuxometerNotification(data)
+	function readLuxometerSuccess(data)
 	{
 		var lux = calculateLux(data)
 		showMessage('Luxometer value: ' + lux)
 	}
 
-	function onLuxometerNotificationError(error)
+	function readLuxometerError(error)
 	{
-		showMessage('Luxometer notification error: ' + error)
+		showMessage('Read Luxometer error: ' + error)
 	}
 }
