+++ examples/deprecated/easyble/tisensortag/easyble.dist.js
