+++ examples/deprecated/easyble/tisensortag/evothings.json
@@ -1,0 +1,3 @@
+{
+  "uuid": "7fffb6b1-736f-4b7a-af9f-86f4d9a45935"
+}
