+++ examples/deprecated/easyble/tisensortag/index.html
@@ -1,0 +1,55 @@
+<!DOCTYPE html>
+<html>
+
+<!--
+Demo of using the EasyBLE library for reading the luxometer of the TI SensorTag CC2650. Please note that EasyBLE is deprecated, instead use the new high-level BLE plugin API.
+-->
+
+<head>
+	<meta charset="utf-8" />
+	<meta name="viewport" content="width=device-width, user-scalable=no,
+		shrink-to-fit=no, initial-scale=1.0, minimum-scale=1.0, maximum-scale=1.0" />
+
+	<title>TI SensorTag Demo - EasyBLE</title>
+
+	<script>
+	// Redirect console.log when running from Evothings Workbench.
+	if (window.hyper && window.hyper.log) { console.log = hyper.log }
+	</script>
+
+	<style>
+	@import "app.css";
+	</style>
+</head>
+
+<body>
+	<img class="evo-image" src="https://evomedia.evothings.com/2013/11/evothings_logo_large_white1.png" />
+
+	<h1>TI SensorTag Demo - EasyBLE</h1>
+
+	<p>This app demonstrates how to use the EasyBLE library for reading the luxometer on the TI SensorTag CC2650.</p>
+
+	<p>Activate the SensorTag. If you don't have a SensorTag you can modify the code in file <code>app.js</code> to work with your device.</p>
+
+	<p id="message">Preparing...</p>
+
+	<div id="sensor-value"></div>
+
+	<!--
+	Including cordova.js will automatically include the JavaScript library
+	for the Cordova BLE plugin.
+	-->
+	<script src="cordova.js"></script>
+
+	<!--
+	Include the EasyBLE library.
+	-->
+	<script src="easyble.dist.js"></script>
+
+	<!--
+	Include application code.
+	-->
+	<script src="app.js"></script>
+</body>
+
+</html>
