--- examples/easyble/tisensortag/evothings.json
@@ -1,3 +1,0 @@
-{
-  "uuid": "1aed2d3c-b619-4fd1-ae71-018925903eaf"
-}
