+++ examples/webbluetooth/tisensortag/evothings.json
@@ -1,3 +1,3 @@
 {
-  "uuid": "2d28f47f-663b-4897-a325-e9ffc473a92a"
+  "uuid": "6f3ea215-0f09-4ab6-903a-4422bcc68c40"
 }
